+++ src/jhv/src/org/helioviewer/jhv/display/RenderListener.java
@@ -2,6 +2,6 @@ package org.helioviewer.jhv.display;
 
 public interface RenderListener {
 
-    public void render(float factor);
+    public void render(double factor);
 
 }
